Override AC_ERLANG_CHECK_LIB m4 macro and use it in configure.ac
This macro is used to get version of an application in erlang instead of getting
it by parsing the lib_dir path.
The versions of Yaws prior to 1.99 does not create versionned path. With this
macro, it is possible to add a requirement on the Yaws version. In this case, we
require Yaws >= 1.90.