Add testsuite support
common_test, eunit and tools application are required to run the testsuites.

With autotools, the results will be placed in the directory
@top_builddir@/testsuite/logs. The coverage result will also be enhanced.

With rebar, the results will be placed in the directory logs in the project root
directory.

For now, only a basic testsuite was added. More tests should be added.